pyproject.toml
[project]
name = "reddit-bigdata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kafka-python>=2.3.0",
    "praw>=7.8.1",
    "prawcore>=2.4.0",
]
